test: refactor and extract cloudwatch test
