+++ packages/artillery/test/cloud-e2e/fargate/fixtures/adot/helpers.js
@@ -69,53 +69,63 @@ async function getDatadogSpans(apiKey, appKey, testId, expectedTotalSpans) {
   return spanList;
 }
 
-async function getXRayTraces(testId) {
+async function getXRayTraces(testId, expectedTraceNum) {
   const endTime = new Date();
   const startTime = new Date(endTime.getTime() - 30 * 60 * 1000); // 30 min ago
   const filterExpression = `annotation.test_id = "${testId}"`;
 
-  let summariesData;
-  try {
-    summariesData = await new Promise((resolve, reject) => {
-      xray.getTraceSummaries(
-        {
+  const maxRetry = 12;
+  const delay = 30000;
+  let retryNum = 0;
+  let traceSummaries = [];
+
+  while (traceSummaries.length < expectedTraceNum && retryNum <= maxRetry) {
+    try {
+      console.log(
+        `ADOT Cloudwatch test: Awaiting trace summaries... (retry #${retryNum})`
+      );
+      traceSummaries = await xray
+        .getTraceSummaries({
           StartTime: startTime,
           EndTime: endTime,
           FilterExpression: filterExpression,
           TimeRangeType: 'Event'
-        },
-        (err, data) => {
-          if (err) reject(err);
-          else resolve(data);
-        }
-      );
-    });
-  } catch (err) {
-    console.error('Error fetching trace summaries from AWS X-Ray: ', err);
+        })
+        .promise()
+        .then((data) => data.TraceSummaries);
+
+      await sleep(delay);
+      retryNum++;
+    } catch (err) {
+      throw new Error(err);
+    }
   }
 
-  const traceSummaries = summariesData.TraceSummaries;
   const traceIds = traceSummaries.map((trace) => trace.Id);
   console.log('TRACE IDS: ', traceIds);
 
-  let fullTraceData;
-  try {
-    fullTraceData = await new Promise((resolve, reject) => {
-      xray.batchGetTraces(
-        {
+  let fullTraceData = [];
+  let retryBatchNum = 0;
+  while (fullTraceData.length < expectedTraceNum && retryNum <= maxRetry) {
+    console.log(
+      `ADOT Cloudwatch test: Awaiting traces... (retry #${retryBatchNum})`
+    );
+    try {
+      fullTraceData = await xray
+        .batchGetTraces({
           TraceIds: traceIds
-        },
-        (err, data) => {
-          if (err) reject(err);
-          else resolve(data);
-        }
-      );
-    });
-  } catch (err) {
-    console.error('Error fetching full trace data from AWS X-Ray: ', err);
+        })
+        .promise()
+        .then((data) => data.Traces);
+
+      await sleep(delay);
+      retryBatchNum++;
+    } catch (err) {
+      throw new Error(err);
+    }
   }
 
-  const traceMap = fullTraceData.Traces.map((trace) =>
+  const traceMap = fullTraceData?.map((trace) =>
     trace.Segments.map((span) => JSON.parse(span.Document))
   );
